// web/portal/src/types/pedido.ts
export interface Pedido {
  Pedido: number;
  OP: number;
  Referencia: string;
  Maquina: string;
  Seccion: string;
  Proceso: string;
  Usuario: string;
  Estado: string;
  Fabricadas: number;
  FechaInicio: string;
  FechaFin: string;
  Duracion: string;
  DuracionMin: number;
  Ayudante: string;
  InfoExtra1: string;
}